refactor: colocate Zustand stores with UI features
- Move edits store to /ui/profile/store.ts
- Move messaging store to /ui/verification/store.ts
- Delete unused swap.ts and thread.ts stores
- Remove /lib/stores/ folder entirely
- Delete computePendingEdits logic (no longer needed)
- Update all imports to new locations
- Update AGENT.md files to reflect new structure

Stores are now colocated with the features that use them,
improving discoverability and separation of concerns.
